+++ src/Casters/BuilderCaster.php
@@ -2,13 +2,13 @@
 
 namespace Glhd\LaravelDumper\Casters;
 
+use Glhd\LaravelDumper\Support\Properties;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Database\Query\Builder as BaseBuilder;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 use SqlFormatter;
-use Symfony\Component\VarDumper\Caster\CutStub;
 use Symfony\Component\VarDumper\Cloner\Stub;
 
 class BuilderCaster extends Caster
@@ -21,32 +21,32 @@ class BuilderCaster extends Caster
 	
 	/**
 	 * @param BaseBuilder|EloquentBuilder|Relation $target
-	 * @param array $properties
+	 * @param \Glhd\LaravelDumper\Support\Properties $properties
 	 * @param \Symfony\Component\VarDumper\Cloner\Stub $stub
 	 * @param bool $is_nested
 	 * @param int $filter
 	 * @return array
 	 */
-	public function cast($target, array $properties, Stub $stub, bool $is_nested, int $filter = 0): array
+	public function cast($target, Properties $properties, Stub $stub, bool $is_nested, int $filter = 0): array
 	{
-		$result = [
-			Key::virtual('sql') => $this->formatSql($target->toSql(), $target->getBindings()),
-			Key::protected('connection') => $target->getConnection(),
-		];
+		$result = new Properties();
+		
+		$result->putVirtual('sql', $this->formatSql($target->toSql(), $target->getBindings()));
+		$result->putProtected('connection', $target->getConnection());
 		
 		if ($target instanceof EloquentBuilder) {
-			$result[Key::protected('model')] = new CutStub($properties[Key::protected('model')]);
-			$result[Key::protected('eagerLoad')] = $properties[Key::protected('eagerLoad')];
+			$result->copyAndCutProtected('model', $properties);
+			$result->copyProtected('eagerLoad', $properties);
 		}
 		
 		if ($target instanceof Relation) {
-			$result[Key::protected('parent')] = new CutStub($properties[Key::protected('parent')]);
-			$result[Key::protected('related')] = new CutStub($properties[Key::protected('related')]);
+			$result->copyAndCutProtected('parent', $properties);
+			$result->copyAndCutProtected('related', $properties);
 		}
 		
-		$stub->cut += (count($properties) - count($result));
+		$result->applyCutsToStub($stub, $properties);
 		
-		return $result;
+		return $result->all();
 	}
 	
 	protected function formatSql($sql, $bindings): string