+++ src/Casters/Caster.php
@@ -2,6 +2,7 @@
 
 namespace Glhd\LaravelDumper\Casters;
 
+use Glhd\LaravelDumper\Support\Properties;
 use Illuminate\Contracts\Foundation\Application;
 use Symfony\Component\VarDumper\Cloner\AbstractCloner;
 use Symfony\Component\VarDumper\Cloner\Stub;
@@ -10,7 +11,7 @@ abstract class Caster
 {
 	public static array $targets = [];
 	
-	protected static bool $enabled = true;
+	public static bool $enabled = true;
 	
 	public static function register(Application $app): void
 	{
@@ -18,8 +19,8 @@ abstract class Caster
 		
 		foreach (static::$targets as $target) {
 			AbstractCloner::$defaultCasters[$target] = static function($target, array $properties, Stub $stub, bool $is_nested, int $filter = 0) {
-				return static::$enabled
-					? app(static::class)->cast($target, $properties, $stub, $is_nested, $filter)
+				return self::$enabled
+					? app(static::class)->cast($target, new Properties($properties), $stub, $is_nested, $filter)
 					: $properties;
 			};
 		}
@@ -27,13 +28,13 @@ abstract class Caster
 	
 	public static function disable(): void
 	{
-		static::$enabled = false;
+		self::$enabled = false;
 	}
 	
 	public static function enable(): void
 	{
-		static::$enabled = true;
+		self::$enabled = true;
 	}
 	
-	abstract public function cast($target, array $properties, Stub $stub, bool $is_nested, int $filter = 0): array;
+	abstract public function cast($target, Properties $properties, Stub $stub, bool $is_nested, int $filter = 0): array;
 }