+++ src/Casters/ResponseCaster.php
@@ -2,9 +2,9 @@
 
 namespace Glhd\LaravelDumper\Casters;
 
+use Glhd\LaravelDumper\Support\Properties;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Symfony\Component\VarDumper\Caster\Caster as BaseCaster;
 use Symfony\Component\VarDumper\Cloner\Stub;
 
 class ResponseCaster extends Caster
@@ -13,18 +13,17 @@ class ResponseCaster extends Caster
 	
 	/**
 	 * @param Request $target
-	 * @param array $properties
+	 * @param \Glhd\LaravelDumper\Support\Properties $properties
 	 * @param \Symfony\Component\VarDumper\Cloner\Stub $stub
 	 * @param bool $is_nested
 	 * @param int $filter
 	 * @return array
 	 */
-	public function cast($target, array $properties, Stub $stub, bool $is_nested, int $filter = 0): array
+	public function cast($target, Properties $properties, Stub $stub, bool $is_nested, int $filter = 0): array
 	{
-		$result = BaseCaster::filter($properties, BaseCaster::EXCLUDE_NULL, [], $filtered);
-		
-		$stub->cut += $filtered;
-		
-		return $result;
+		return $properties
+			->filter()
+			->applyCutsToStub($stub, $properties)
+			->all();
 	}
 }